% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anc1}
\alias{anc1}
\title{Raw antenatal care dataset for currently pregnant women from the Myanmar
MCCT Programme Evaluation Survey}
\format{A data.frame with 71 columns and 13130 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
anc1
}
\description{
Raw antenatal care dataset for currently pregnant women from the Myanmar
MCCT Programme Evaluation Survey
}
\keyword{datasets}
