% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anc2}
\alias{anc2}
\title{Raw antenatal care dataset for currently non-pregnant women from the Myanmar
MCCT Programme Evaluation Survey}
\format{A data.frame with 183 columns and 11865 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
anc2
}
\description{
Raw antenatal care dataset for currently non-pregnant women from the Myanmar
MCCT Programme Evaluation Survey
}
\keyword{datasets}
