% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ancResults}
\alias{ancResults}
\title{ANC indicators results table}
\format{A data.frame with 9 columns and 918 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
ancResults
}
\description{
ANC indicators results table
}
\keyword{datasets}
