% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anthroChecks}
\alias{anthroChecks}
\title{Anthropometric data checks for Myanmar MCCT Programme Evaluation Survey
household data}
\format{A data.frame with 3 columns and 3043 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
anthroChecks
}
\description{
Anthropometric data checks for Myanmar MCCT Programme Evaluation Survey
household data
}
\keyword{datasets}
