% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anthroDF}
\alias{anthroDF}
\title{Raw anthropometric identifiers dataset from the Myanmar MCCT Programme
Evaluation Survey}
\format{A data.frame with 66 columns and 2670 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
anthroDF
}
\description{
Raw anthropometric identifiers dataset from the Myanmar MCCT Programme
Evaluation Survey
}
\keyword{datasets}
