% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birthResults}
\alias{birthResults}
\title{Birth/delivery indicators results table}
\format{A data.frame with 9 columns and 486 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
birthResults
}
\description{
Birth/delivery indicators results table
}
\keyword{datasets}
