% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccsiResults}
\alias{ccsiResults}
\title{CCSI indicators results table}
\format{A data.frame with 9 columns and 18 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
ccsiResults
}
\description{
CCSI indicators results table
}
\keyword{datasets}
