% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{checks}
\alias{checks}
\title{Data checks for Myanmar MCCT Programme Evaluation Survey household data}
\format{A data.frame with 3 columns and 1028 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
checks
}
\description{
Data checks for Myanmar MCCT Programme Evaluation Survey household data
}
\keyword{datasets}
