% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{childAnthro}
\alias{childAnthro}
\title{Raw child anthropometric dataset from the Myanmar MCCT Programme Evaluation
Survey}
\format{A data.frame with 166 columns and 3136 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
childAnthro
}
\description{
Raw child anthropometric dataset from the Myanmar MCCT Programme Evaluation
Survey
}
\keyword{datasets}
