% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{childHealth}
\alias{childHealth}
\title{Raw child health dataset from the Myanmar MCCT Programme Evaluation Survey}
\format{A data.frame with 126 columns and 13130 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
childHealth
}
\description{
Raw child health dataset from the Myanmar MCCT Programme Evaluation Survey
}
\keyword{datasets}
