% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-clean_hh.R
\name{clean_anthro}
\alias{clean_anthro}
\title{Function to clean anthropometric data}
\usage{
clean_anthro(df, checks)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing anthropometric data.}

\item{checks}{A data.frame containing check values for cleaning.}
}
\value{
A data.frame equivalent to \code{df} but with extraneous observations
  and other variables cleaned ready for further processing.
}
\description{
Function to clean anthropometric data
}
\examples{
\dontrun{
clean_anthro(df = anthroData[[1]], checks = checks)
}

}
