% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-clean_hh.R
\name{clean_hh}
\alias{clean_hh}
\title{Function to clean household data}
\usage{
clean_hh(df, checks)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing household data. Default is \code{hh}.}

\item{checks}{A data.frame containing check values for cleaning.}
}
\value{
A data.frame equivalent to \code{df} but with extraneous observations
  and other variables cleaned ready for further processing.
}
\description{
Function to clean household data
}
\examples{
clean_hh(df = hh, checks = checks)

}
