% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnutResults}
\alias{cnutResults}
\title{Child nutrition indicators results table}
\format{A data.frame with 9 columns and 128 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
cnutResults
}
\description{
Child nutrition indicators results table
}
\keyword{datasets}
