% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codebook}
\alias{codebook}
\title{Codebook for Myanmar MCCT Programme Evaluation Survey main dataset}
\format{A data.frame in \code{tbl} format with 3 columns and 833 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
codebook
}
\description{
Codebook for Myanmar MCCT Programme Evaluation Survey main dataset
}
\keyword{datasets}
