% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codebookAnthro}
\alias{codebookAnthro}
\title{Codebook for Myanmar MCCT Programme Evaluation Survey anthropometric dataset}
\format{A data.frame in \code{tbl} format with 3 columns and 162 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
codebookAnthro
}
\description{
Codebook for Myanmar MCCT Programme Evaluation Survey anthropometric dataset
}
\keyword{datasets}
