% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codebookTownship}
\alias{codebookTownship}
\title{Codebook for Myanmar MCCT Programme Evaluation Survey township profile
dataset}
\format{A data.frame in \code{tbl} format with 3 columns and 64 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
codebookTownship
}
\description{
Codebook for Myanmar MCCT Programme Evaluation Survey township profile
dataset
}
\keyword{datasets}
