% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codebookVillage}
\alias{codebookVillage}
\title{Codebook for Myanmar MCCT Programme Evaluation Survey village profile
dataset}
\format{A data.frame in \code{tbl} format with 3 columns and 432 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
codebookVillage
}
\description{
Codebook for Myanmar MCCT Programme Evaluation Survey village profile
dataset
}
\keyword{datasets}
