% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-recode_anc.R
\name{create_anc}
\alias{create_anc}
\title{Function to create raw ANC dataset}
\usage{
create_anc(df, x, y, status)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing ANC data}

\item{x}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing household information that will be linked to the
ANC data}

\item{y}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing household roster information that will be linked to the
ANC data}

\item{status}{A character value indicating whether to create data of
currently pregnant women ("current") or previously pregnant women ("past")}
}
\value{
A data.frame containing raw ANC data for the Myanmar MCCT Programme
  Evaluation Study
}
\description{
Function to create raw ANC dataset
}
\examples{
create_anc(df = anc1, x = hh, y = hhMembers, status = "current")
create_anc(df = anc2, x = hh, y = hhMembers, status = "past")

}
