% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/24-create_tables.R
\name{create_anthro_table}
\alias{create_anthro_table}
\title{Function to create indicator results tables}
\usage{
create_anthro_table(df, vars, digits = NULL)
}
\arguments{
\item{df}{A recoded data.frame of indicators}

\item{vars}{Indicator/s to create a results table for}

\item{digits}{Integer value for number of digits to round off values.
Default is NULL for no rounding off.}
}
\value{
A data.frame of indicator results for each state in total and
  disaggregated by urban, rural and hard-to-reach areas and by wealth
  quintile.
}
\description{
Function to create indicator results tables
}
\examples{
## Recode wash indicators
canthro  <- recode_anthro(df = create_canthro(df = childAnthro, x = anthroDF))

create_anthro_table(df = canthro,
                    vars = c("haz", "global.haz", "moderate.haz", "severe.haz"))

}
