% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-recode_anthro.R
\name{create_canthro}
\alias{create_canthro}
\title{Function to create raw child antropometry dataset}
\usage{
create_canthro(df, x)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing child anthropometry data}

\item{x}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing identifying information that will be linked to the
child anthropometry data}
}
\value{
A data.frame containing raw child anthropometry data for the Myanmar
  MCCT Programme Evaluation Study
}
\description{
Function to create raw child antropometry dataset
}
\examples{
create_canthro(df = childAnthro, x = anthroDF)

}
