% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-recode_chealth.R
\name{create_chealth}
\alias{create_chealth}
\title{Function to create raw child health dataset}
\usage{
create_chealth(df, x, y)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing child health data}

\item{x}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing household information that will be linked to the
child health data}

\item{y}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing household roster information that will be linked to the
child health data}
}
\value{
A data.frame containing raw child health data for the Myanmar
  MCCT Programme Evaluation Study
}
\description{
Function to create raw child health dataset
}
\examples{
create_chealth(df = childHealth, x = hh, y = hhMembers)

}
