% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-recode_anthro.R
\name{create_manthro}
\alias{create_manthro}
\title{Function to create raw mother antropometry dataset}
\usage{
create_manthro(df, x, y)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing mother anthropometry data}

\item{x}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing identifying information that will be linked to the
mother anthropometry data}

\item{y}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing child anthropometry data to get associated information
that will be linked to the mother anthropometry data}
}
\value{
A data.frame containing raw mother anthropometry data for the Myanmar
  MCCT Programme Evaluation Study
}
\description{
Function to create raw mother antropometry dataset
}
\examples{
create_manthro(df = motherAnthro, x = anthroDF,
               y = create_canthro(df = childAnthro, x = anthroDF))

}
