% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19-recode_mcct.R
\name{create_mcct}
\alias{create_mcct}
\title{Function to create dataset with individual-level MCCT-related data from the
Myanmar MCCT Programme Evaluation survey}
\usage{
create_mcct(df, x)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing MCCT-related data}

\item{x}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing household member information that will be linked to the
MCCT data}
}
\value{
A data.frame of recoded MCCT-related indicators for children and
  pregnant women
}
\description{
Function to create dataset with individual-level MCCT-related data from the
Myanmar MCCT Programme Evaluation survey
}
\examples{
## Recode individual MCCT-related indicators
create_mcct(df = hh, x = hhMembers)

}
