% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19-recode_mcct.R
\name{create_mcct_household}
\alias{create_mcct_household}
\title{Function to create dataset with household-level MCCT-related data from the
Myanmar MCCT Programme Evaluation survey}
\usage{
create_mcct_household(df, x)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing houehold-level MCCT-related data}

\item{x}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing household member information that will be linked to the
MCCT data}
}
\value{
A data.frame of recoded MCCT-related indicators for households
}
\description{
Function to create dataset with household-level MCCT-related data from the
Myanmar MCCT Programme Evaluation survey
}
\examples{
## Recode household MCCT-related indicators
create_mcct_household(df = hh, x = hhMembers)

}
