% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-recode_ppi.R
\name{create_ppi}
\alias{create_ppi}
\title{Function to create poverty probability index dataset for the Myanmar MCCT
Programme Evaluation survey}
\usage{
create_ppi(df)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing poverty-related household data}
}
\value{
A data.frame of poverty probability index dataset
}
\description{
Function to create poverty probability index dataset for the Myanmar MCCT
Programme Evaluation survey
}
\examples{
## Create poverty probability index dataset
create_ppi(df = hh)

}
