% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/24-create_tables.R
\name{create_table}
\alias{create_table}
\title{Function to create indicator results tables}
\usage{
create_table(df, var)
}
\arguments{
\item{df}{A recoded data.frame of indicators}

\item{var}{Indicator to create a results table for}
}
\value{
A data.frame of indicator results with 95\% confidence limits for
  each state in total and disaggregated by urban, rural and hard-to-reach
  areas.
}
\description{
This function performs an exact binomial test on the data for the indicator
of interest and provides an estimate of the indicator proportion and the
corresponding 95\% confidence interval. This assumes that the probability
weights of selection of an respondent are equal (given that sample has
been taken using probability proportional to population size, this
assumption holds true unless otherwise stated).
}
\examples{
## Recode anthro indicators
temp <- recode_anthro(df = create_canthro(df = childAnthro, x = anthroDF))

create_table(df = temp, var = "global.haz")

}
