% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/24-create_tables.R
\name{create_unweighted_table}
\alias{create_unweighted_table}
\title{Function to create indicator results tables}
\usage{
create_unweighted_table(df, ppiDF = recode_ppi(df = hh), vars, digits = NULL)
}
\arguments{
\item{df}{A recoded data.frame of indicators}

\item{ppiDF}{A data.frame produced when \code{recode_ppi} is applied to
the \code{hh} dataset of the Myanmar MCCT Programme Evaluation}

\item{vars}{Indicator/s to create a results table for}

\item{digits}{Integer value for number of digits to round off values.
Default is NULL for no rounding off.}
}
\value{
A data.frame of indicator results for each state in total and
  disaggregated by urban, rural and hard-to-reach areas and by wealth
  quintile.
}
\description{
Function to create indicator results tables
}
\examples{
## Recode wash indicators
washDF  <- recode_wash(df = hh)

create_unweighted_table(df = washDF,
                        vars = c("summer1", "summer6", "summer5",
                                 "rain1", "rain6", "rain5",
                                 "winter1", "winter6", "winter5"))

}
