% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/24-create_tables.R
\name{create_weighted_anthro}
\alias{create_weighted_anthro}
\title{Function to create weighted anthropometric indicator results tables}
\usage{
create_weighted_anthro(df, sex = FALSE, vars, labs, state, digits = NULL)
}
\arguments{
\item{df}{A recoded data.frame of indicators}

\item{sex}{Logical. Should results be disaggregated by sex? Default is FALSE.}

\item{vars}{Variable names for anthropometric indicator/s to create a results
table for}

\item{labs}{Variable labels for anthropometric indicator/s to create a
results table for}

\item{state}{A character value for name of state to report results on.
Choice between Kayah or Kayin.}

\item{digits}{Integer value for number of digits to round off values.
Default is NULL for no rounding off.}
}
\value{
A data.frame of anthropometric indicator results for each state in
  total and disaggregated by urban, rural and hard-to-reach areas
}
\description{
Function to create weighted anthropometric indicator results tables
}
\examples{
## Recode stunting indicators
canthro  <- recode_anthro(df = create_canthro(df = childAnthro, x = anthroDF))

create_weighted_anthro(df = canthro,
                       vars = c("haz", "global.haz", "moderate.haz", "severe.haz"),
                       labs = c("Mean height-for-age z-score",
                                "Global stunting/stuntedness",
                                "Moderate stunting/stuntedness",
                                "Severe stunting/stuntedness"),
                       state = "Kayah")

create_weighted_anthro(df = canthro, sex = TRUE,
                       vars = c("haz", "global.haz", "moderate.haz", "severe.haz"),
                       labs = c("Mean height-for-age z-score",
                                "Global stunting/stuntedness",
                                "Moderate stunting/stuntedness",
                                "Severe stunting/stuntedness"),
                       state = "Kayah")

}
