% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/24-create_tables.R
\name{create_weighted_table}
\alias{create_weighted_table}
\title{Function to create weighted indicator results tables}
\usage{
create_weighted_table(
  df,
  sex = FALSE,
  ppiDF = recode_ppi(df = hh),
  vars,
  labs,
  state,
  digits = NULL
)
}
\arguments{
\item{df}{A recoded data.frame of indicators}

\item{sex}{Logical. Should results be disaggregated by sex? Default is FALSE}

\item{ppiDF}{A data.frame produced when \code{recode_ppi} is applied to
the \code{hh} dataset of the Myanmar MCCT Programme Evaluation}

\item{vars}{Variable names for indicator/s to create a results table for}

\item{labs}{Variable labels for indicator/s to create a results table for}

\item{state}{A character value for name of state to report results on.
Choice between Kayah or Kayin.}

\item{digits}{Integer value for number of digits to round off values.
Default is NULL for no rounding off.}
}
\value{
A data.frame of indicator results for each state in total and
  disaggregated by urban, rural and hard-to-reach areas and by wealth
  quintile.
}
\description{
Function to create weighted indicator results tables
}
\examples{
## Recode MCCT indicators
mcctDF  <- recode_mcct(df = create_mcct(df = hh, x = hhMembers))

create_weighted_table(df = mcctDF,
                      vars = c("mcct1", "mcct2"),
                      labs = c("MCCT coverage",
                               "Mean number of cash transfers received"),
                      state = "Kayah")

## Recode IYCF indicators
iycfDF <- recode_iycf(df = create_iycf(df = iycf, x = hh, y = hhMembers))

create_weighted_table(df = iycfDF,
                      sex = TRUE,
                      vars = c("eibf", "ebf", "mmf1", "mmf2", "mmf3", "mmf",
                               "fgscore", "mdd", "mad1", "mad2", "mad"),
                      labs = c("Early initiation of breastfeeding",
                               "Exclusive breastfeeding",
                               "Minimum meal frequency for breastfed 6-8 month old",
                               "Minimum meal frequency for breastfed 9-23 month old",
                               "Minimum meal frequency for non-breastfed 6-23 month old",
                               "Minimum meal frequency",
                               "Mean number of food groups consumed",
                               "Minimum dietary diversity",
                               "Minimum adequate diet for breastfed children",
                               "Minimum adequate diet for non-breastfed children",
                               "Minimum adequate diet"),
                      state = "Kayah")


}
