% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/26-township.R
\name{describe_township}
\alias{describe_township}
\title{Function to describe township}
\usage{
describe_township(df, state = c("Kayah", "Kayin"))
}
\arguments{
\item{df}{Dataset containing data collected from townships included in
the MCCT baseline study}

\item{state}{State for which township description is being generated for}
}
\value{
A character vector describing each township
}
\description{
Function to describe township
}
\examples{
describe_township(df = recode_township(df = townshipData),
                  state = "Kayah")

}
