% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/27-village.R
\name{describe_village}
\alias{describe_village}
\title{Function to create a description and summary of village data information}
\usage{
describe_village(df, state = c("Kayah", "Kayin"))
}
\arguments{
\item{df}{A data.frame with recoded village data information for analysis}

\item{state}{State for which to produce analysis for.}
}
\value{
A list of descriptions and summaries.
}
\description{
Function to create a description and summary of village data information
}
\examples{
describe_village(df = recode_village(df = villageData1), state = "Kayah")

}
