% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diarrhoeaResults}
\alias{diarrhoeaResults}
\title{Diarrhoea treatment indicators results table}
\format{A data.frame with 9 columns and 540 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
diarrhoeaResults
}
\description{
Diarrhoea treatment indicators results table
}
\keyword{datasets}
