% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fcsnResults}
\alias{fcsnResults}
\title{FCS-N indicators results table}
\format{A data.frame with 9 columns and 162 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
fcsnResults
}
\description{
FCS-N indicators results table
}
\keyword{datasets}
