% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fpResults}
\alias{fpResults}
\title{Family planning indicators results table}
\format{A data.frame with 9 columns and 774 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
fpResults
}
\description{
Family planning indicators results table
}
\keyword{datasets}
