% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/23-calculate_weights.R
\name{get_cluster_weights}
\alias{get_cluster_weights}
\title{Function to calculate probability of each cluster in a sampling frame to be
sampled}
\usage{
get_cluster_weights(n, N, m)
}
\arguments{
\item{n}{Population size of cluster}

\item{N}{Total population}

\item{m}{Number of clusters}
}
\value{
Numeric value of cluster weight
}
\description{
Function to calculate probability of each cluster in a sampling frame to be
sampled
}
\examples{
get_cluster_weights(n = 1028, N = 20000, m = 30)

}
