% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/23-calculate_weights.R
\name{get_individual_weights}
\alias{get_individual_weights}
\title{Function to calculate probability of an individual being sampled in a
cluster}
\usage{
get_individual_weights(n, c)
}
\arguments{
\item{n}{Population size of cluster}

\item{c}{Number of individuals sampled in a cluster}
}
\value{
Numeric value of probability of an individual being sampled in a
cluster
}
\description{
Function to calculate probability of an individual being sampled in a
cluster
}
\examples{
get_individual_weights(n = 1028, c = 300)

}
