% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-get_data.R
\name{get_mcct_data}
\alias{get_mcct_data}
\title{Function to get Myanmar MCCT study datasets and forms from ODK server.}
\usage{
get_mcct_data(
  id,
  username,
  password,
  start = Sys.Date(),
  end = Sys.Date(),
  filename,
  rep = FALSE,
  rep.name = NULL
)
}
\arguments{
\item{id}{Form ID. Can be one of several form IDs used by the
different forms designed for the MCCT study -
\code{baseline_mcct_final}, \code{baseline_mcct_anthro_final},
\code{baseline_townshipprofile_final}, \code{baseline_villprofile_final},
\code{MCCT_Baseline_Spotcheck}.}

\item{username}{ONA server username credentials. For}

\item{password}{ONA server password credentials.}

\item{start}{Include data from submission dates after (inclusive) this
start date in export to CSV. Date format \code{<yyyy/MM/dd>} and default is
current system date}

\item{end}{Include data from submission dates before (exclusive) this date
in export to CSV. Date format is \code{<yyyy/MM/dd>} and default value current
system date}

\item{filename}{Filename to use for data without the CSV file extension.}

\item{rep}{Logical. Does the form have repeat/s? Default FALSE.}

\item{rep.name}{A vector of repeat names to read in the form. Default
is NULL. Must be specified if \code{rep} is TRUE.}
}
\value{
A data.frame corresponding to dataset corresponding to the form ID
  specified.
}
\description{
This function is a wrapper for a set of functions from the \code{odkr}
package.
}
\examples{
\dontrun{
  get_mcct_data(id = "baseline_mcct_final",
                username = "validmeasures",
                password = "6Y2-8yK-Nmk-Lbf",
                filename = "main_form")
}

}
