% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/29-get_n.R
\name{get_n}
\alias{get_n}
\title{Function to get number of samples used to calculate indicator}
\usage{
get_n(df, sex = FALSE, ppiDF = recode_ppi(df = hh), vars)
}
\arguments{
\item{df}{A data.frame of recoded indicator variables}

\item{sex}{Logical. Should sample size by sex be reported? Defualt is FALSE}

\item{ppiDF}{A data.frame produced when \code{recode_ppi} is applied to
the \code{hh} dataset of the Myanmar MCCT Programme Evaluation}

\item{vars}{A character vector of variable names to get number of samples for}
}
\value{
A table of numeric values of number of samples use to calculate
  specific indicators by area and wealth strata
}
\description{
Function to get number of samples used to calculate indicator
}
\examples{

get_n(df = recode_mcct(create_mcct(df = hh, x = hhMembers)),
      vars = c("mcct1", "mcct2"))

}
