% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/29-get_n.R
\name{get_n_area}
\alias{get_n_area}
\title{Function to get number of samples used to calculate indicator stratified
by geographic area}
\usage{
get_n_area(df, vars)
}
\arguments{
\item{df}{A data.frame of recoded indicator variables}

\item{vars}{A character vector of variable names to get number of samples for}
}
\value{
A table of numeric values of number of samples use to calculate
  specific indicators by area strata
}
\description{
Function to get number of samples used to calculate indicator stratified
by geographic area
}
\examples{

get_n_area(df = recode_anthro(df = create_canthro(df = childAnthro,
                                                  x = anthroDF)),
           vars = c("global.haz", "moderate.haz", "severe.haz"))

}
