% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/29-get_n.R
\name{get_n_wealth}
\alias{get_n_wealth}
\title{Function to get number of samples used to calculate indicator stratified
by wealth}
\usage{
get_n_wealth(df, ppiDF = recode_ppi(df = hh), vars)
}
\arguments{
\item{df}{A data.frame of recoded indicator variables}

\item{ppiDF}{A data.frame produced when \code{recode_ppi} is applied to
the \code{hh} dataset of the Myanmar MCCT Programme Evaluation}

\item{vars}{A character vector of variable names to get number of samples for}
}
\value{
A table of numeric values of number of samples use to calculate
  specific indicators by wealth strata
}
\description{
Function to get number of samples used to calculate indicator stratified
by wealth
}
\examples{

get_n_wealth(df = recode_mcct(create_mcct(df = hh, x = hhMembers)),
             vars = c("mcct1", "mcct2"))

}
