% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/23-calculate_weights.R
\name{get_weights}
\alias{get_weights}
\title{Function to calculate the inverse of the probability of an individual being
sampled in the population}
\usage{
get_weights(n, N, m, c)
}
\arguments{
\item{n}{Population size of cluster}

\item{N}{Total population}

\item{m}{Number of clusters}

\item{c}{Number of individuals to be sampled in a cluster}
}
\value{
Numeric value of inverse probability weight
}
\description{
Function to calculate the inverse of the probability of an individual being
sampled in the population
}
\examples{
get_weights(n = 1028, N = 20000, m = 30, c = 300)

}
