% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hfesResults}
\alias{hfesResults}
\title{HFES indicators results table}
\format{A data.frame with 9 columns and 72 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
hfesResults
}
\description{
HFES indicators results table
}
\keyword{datasets}
