% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hh}
\alias{hh}
\title{Raw household dataset from the Myanmar MCCT Programme Evaluation Survey}
\format{A data.frame with 538 columns and 2393 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
hh
}
\description{
Raw household dataset from the Myanmar MCCT Programme Evaluation Survey
}
\keyword{datasets}
