% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hhMembers}
\alias{hhMembers}
\title{Raw household members dataset from the Myanmar MCCT Programme Evaluation
Survey}
\format{A data.frame with 49 columns and 13097 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
hhMembers
}
\description{
Raw household members dataset from the Myanmar MCCT Programme Evaluation
Survey
}
\keyword{datasets}
