% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{illResults}
\alias{illResults}
\title{Childhood illness indicators results table}
\format{A data.frame with 9 columns and 54 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
illResults
}
\description{
Childhood illness indicators results table
}
\keyword{datasets}
