% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{immunisationResults}
\alias{immunisationResults}
\title{Immunisation indicators results table}
\format{A data.frame with 9 columns and 378 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
immunisationResults
}
\description{
Immunisation indicators results table
}
\keyword{datasets}
