% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iycf}
\alias{iycf}
\title{Raw IYCF dataset from the Myanmar MCCT Programme Evaluation Survey}
\format{A data.frame with 71 columns and 13272 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
iycf
}
\description{
Raw IYCF dataset from the Myanmar MCCT Programme Evaluation Survey
}
\keyword{datasets}
