% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcctResults}
\alias{mcctResults}
\title{MCCT indicators results table}
\format{A data.frame with 9 columns and 288 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
mcctResults
}
\description{
MCCT indicators results table
}
\keyword{datasets}
