% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mddwResults}
\alias{mddwResults}
\title{MDDW indicators results table}
\format{A data.frame with 9 columns and 18 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
mddwResults
}
\description{
MDDW indicators results table
}
\keyword{datasets}
