% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnutResults}
\alias{mnutResults}
\title{Maternal nutrition indicators results table}
\format{A data.frame with 9 columns and 32 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
mnutResults
}
\description{
Maternal nutrition indicators results table
}
\keyword{datasets}
