% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motherAnthro}
\alias{motherAnthro}
\title{Raw mother anthropometric dataset from the Myanmar MCCT Programme Evaluation
Survey}
\format{A data.frame with 75 columns and 2478 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
motherAnthro
}
\description{
Raw mother anthropometric dataset from the Myanmar MCCT Programme Evaluation
Survey
}
\keyword{datasets}
