% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myanmarMCCTdata.R
\docType{package}
\name{myanmarMCCTdata}
\alias{myanmarMCCTdata}
\alias{myanmarMCCTdata-package}
\title{myanmarMCCTdata: Myanmar Mother and Child Cash Transfer (MCCT) Programme Evaluation Data
    Handler}
\description{
In support of the Myanmar Mother and Child Cash Transfer programme 
    evaluation, this package has been developed to support the handling,
    processing, analysis of data collected for the study. The package includes 
    functions to retrieve data from the online server database, to appropriately
    structure datasets, to check and clean data, to recode data to respective
    indicator sets, to estimate indicators and to perform appropriate
    comparative analysis specified by the analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/validmeasures/myanmarMCCTdata}
  \item Report bugs at \url{https://github.com/validmeasures/myanmarMCCTdata/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

Authors:
\itemize{
  \item Abdoul-Aziz Goza \email{aziz@validinternational.org}
}

Other contributors:
\itemize{
  \item Nicholus Tint Zaw \email{nicholustintzaw@gmail.com} [contributor]
}

}
\keyword{internal}
