% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nbcResults}
\alias{nbcResults}
\title{NBC indicators results table}
\format{A data.frame with 9 columns and 540 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
nbcResults
}
\description{
NBC indicators results table
}
\keyword{datasets}
