% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pncResults}
\alias{pncResults}
\title{PNC indicators results table}
\format{A data.frame with 9 columns and 324 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
pncResults
}
\description{
PNC indicators results table
}
\keyword{datasets}
