% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-recode_anthro.R
\name{recode_age}
\alias{recode_age}
\title{Function to recode age variable and convert to age in days}
\usage{
recode_age(df)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing anthropometric data}
}
\value{
A vector of recoded ages based on any re-testing done
}
\description{
Function to recode age variable and convert to age in days
}
\examples{

recode_age(df = childAnthro)

}
