% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-recode_anc.R
\name{recode_anc}
\alias{recode_anc}
\title{Function to recode antenatal care indicators for the Myanmar MCCT Programme
Evaluation Survey}
\usage{
recode_anc(
  df,
  status = NULL,
  core.columns = c("KEY", "geo_rural", "geo_state", "geo_villward", "sample_component")
)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing antenatal care data}

\item{status}{A character value indicating whether to recode data of
currently pregnant women ("current") or previously pregnant women ("past")}

\item{core.columns}{A vector of variable names to include in resulting}
}
\value{
A data.frame of recoded antenatal care indicators
}
\description{
Function to recode antenatal care indicators for the Myanmar MCCT Programme
Evaluation Survey
}
\examples{
## Recode anc indicators for currently pregnant women
x <- create_anc(df = anc1, x = hh, y = hhMembers, status = "current")
recode_anc(df = x, status = "current")

## Recode anc indicators for non-pregnant women
x <- create_anc(df = anc2, x = hh, y = hhMembers, status = "past")
recode_anc(df = x, status = "past")

}
