% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-recode_anc.R
\name{recode_anc_supplementation}
\alias{recode_anc_supplementation}
\title{Function to recode supplementation indicators for antenatal care for the
Myanmar MCCT Programme Evaluation Survey}
\usage{
recode_anc_supplementation(df, status = NULL)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing antenatal care data}

\item{status}{A character value indicating whether to recode data of
currently pregnant women ("current") or previously pregnant women ("past")}
}
\value{
A data.frame of recoded supplementation indicators associated with
  antenatal care
}
\description{
Function to recode supplementation indicators for antenatal care for the
Myanmar MCCT Programme Evaluation Survey
}
\examples{
## Recode anc supplementation for currently pregnant women
x <- create_anc(df = anc1, x = hh, y = hhMembers, status = "current")
recode_anc_supplementation(df = x, status = "current")

## Recode anc supplementation for non-pregnant women
x <- create_anc(df = anc2, x = hh, y = hhMembers, status = "past")
recode_anc_supplementation(df = x, status = "past")

}
