% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-recode_anc.R
\name{recode_anc_testing}
\alias{recode_anc_testing}
\title{Function to recode testing for infections indicators for antenatal care for
the Myanmar MCCT Programme Evaluation Survey}
\usage{
recode_anc_testing(df, status = NULL)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing antenatal care data}

\item{status}{A character value indicating whether to recode data of
currently pregnant women ("current") or previously pregnant women ("past")}
}
\value{
A data.frame of recoded testing for infections associated with
  antenatal care
}
\description{
Function to recode testing for infections indicators for antenatal care for
the Myanmar MCCT Programme Evaluation Survey
}
\examples{
## Recode data for currently pregnant women
x <- create_anc(df = anc1, x = hh, y = hhMembers, status = "current")
recode_anc_testing(df = x, status = "current")

## Recode data for non-pregnant women
x <- create_anc(df = anc2, x = hh, y = hhMembers, status = "past")
recode_anc_testing(df = x, status = "past")

}
