% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-recode_anthro.R
\name{recode_anthro}
\alias{recode_anthro}
\title{Function to recode child anthropometric indicators and calculate
anthropometric z-scores}
\usage{
recode_anthro(
  df,
  core.columns = c("enumDate", "geo_rural", "geo_state", "geo_villward"),
  flag = TRUE,
  cases = TRUE
)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing anthropometric data}

\item{core.columns}{A vector of variable names to include in resulting}

\item{flag}{Logical. Should WHO flagging criteria be applied to the z-scores?
Default is TRUE.}

\item{cases}{Logical. Should cases of child undernutrition be assessed?
Default is TRUE.}
}
\value{
A data.frame of recoded anthropometric measurements and z-scores
}
\description{
Function to recode child anthropometric indicators and calculate
anthropometric z-scores
}
\examples{
recode_anthro(df = create_canthro(df = childAnthro, x = anthroDF))

}
