% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-recode_chealth.R
\name{recode_ari}
\alias{recode_ari}
\title{Function to recode cough/ARI prevalence and treatment-seeking behaviour for
the Myanmar MCCT Programme Evaluation Survey}
\usage{
recode_ari(df)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing child health data}
}
\value{
A vector of recoded cough/ARI prevalence and treatment-seeking
  behaviour
}
\description{
Function to recode cough/ARI prevalence and treatment-seeking behaviour for
the Myanmar MCCT Programme Evaluation Survey
}
\examples{
chealth <- create_chealth(df = childHealth, x = hh, y = hhMembers)
recode_ari(df = chealth)

}
