% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-recode_delivery.R
\name{recode_birth}
\alias{recode_birth}
\title{Function to recode the birth/delivery indicators for the Myanmar MCCT
Programme Evaluation survey}
\usage{
recode_birth(
  df,
  core.columns = c("KEY", "geo_rural", "geo_state", "geo_villward", "sample_component")
)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing birth/delivery data}

\item{core.columns}{A vector of variable names to include in resulting
data.frame}
}
\value{
A data.frame of recoded birth/delivery indicators
}
\description{
Function to recode the birth/delivery indicators for the Myanmar MCCT
Programme Evaluation survey
}
\examples{
## Recode birth/delivery indicators
x <- create_anc(df = anc2, x = hh, y = hhMembers, status = "past")
recode_birth(df = x)

}
