% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-recode_chealth.R
\name{recode_birthweight}
\alias{recode_birthweight}
\title{Function to recode birth weight indicators for the Myanmar MCCT
Programme Evaluation Survey}
\usage{
recode_birthweight(df)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing child health data}
}
\value{
A vector of recoded birth weight in grams
}
\description{
Function to recode birth weight indicators for the Myanmar MCCT
Programme Evaluation Survey
}
\examples{
chealth <- create_chealth(df = childHealth, x = hh, y = hhMembers)
recode_birthweight(df = chealth)

}
