% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-recode_iycf.R
\name{recode_comp_feed}
\alias{recode_comp_feed}
\title{Function to recode timely complementary feeding}
\usage{
recode_comp_feed(df)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing IYCF/breastfeeding data}
}
\value{
A vector of recoded timely complementary feeding status
}
\description{
Function to recode timely complementary feeding
}
\examples{
iycfDF <- create_iycf(df = iycf, x = hh, y = hhMembers)
recode_comp_feed(df = iycfDF)

}
