% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-recode_iycf.R
\name{recode_cont_bf}
\alias{recode_cont_bf}
\title{Function to recode continued breastfeeding indicators}
\usage{
recode_cont_bf(df)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing IYCF/breastfeeding data}
}
\value{
A vector of recoded continued breastfeeding status
}
\description{
Function to recode continued breastfeeding indicators
}
\examples{
iycfDF <- create_iycf(df = iycf, x = hh, y = hhMembers)
recode_cont_bf(df = iycfDF)

}
