% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-recode_csi.R
\name{recode_csi_consumption}
\alias{recode_csi_consumption}
\title{Function to recode consumption-based Coping Strategies Index for the Myanmar
MCCT Programme Evaluation Survey}
\usage{
recode_csi_consumption(
  df,
  core.columns = c("KEY", "geo_state", "geo_rural", "geo_villward", "sample_component")
)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing consumption-based coping stratgies index data}

\item{core.columns}{A vector of variable names to include in resulting
data.frame}
}
\value{
A vector of recoded consumption-based coping strategies index
  indicators
}
\description{
Function to recode consumption-based Coping Strategies Index for the Myanmar
MCCT Programme Evaluation Survey
}
\examples{
recode_csi_consumption(df = hh)

}
