% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-recode_iycf.R
\name{recode_eibf}
\alias{recode_eibf}
\title{Function to recode early initiation of breastfeeding indicator for the
Myanmar MCCT Programme Evaluation Survey}
\usage{
recode_eibf(df)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing IYCF/breastfeeding data}
}
\value{
A vector of recoded early initiation of breastfeeding status
}
\description{
Function to recode early initiation of breastfeeding indicator for the
Myanmar MCCT Programme Evaluation Survey
}
\examples{
iycfDF <- create_iycf(df = iycf, x = hh, y = hhMembers)
recode_eibf(df = iycfDF)

}
