% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-recode_chealth.R
\name{recode_epi}
\alias{recode_epi}
\title{Function to recode vaccine coverage indicators for the Myanmar MCCT
Programme Evaluation Survey}
\usage{
recode_epi(df)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing child health data}
}
\value{
A vector of recoded vaccination coverage indicators
}
\description{
Function to recode vaccine coverage indicators for the Myanmar MCCT
Programme Evaluation Survey
}
\examples{
chealth <- create_chealth(df = childHealth, x = hh, y = hhMembers)
recode_epi(df = chealth)

}
