% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-recode_fcs.R
\name{recode_fcs}
\alias{recode_fcs}
\title{Function to recode Food Consumption Score for the Myanmar MCCT
Programme Evaluation Survey}
\usage{
recode_fcs(
  df,
  x = hhMembers,
  core.columns = c("KEY", "geo_state", "geo_rural", "geo_villward", "sample_component",
    "hh_mem_sex")
)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing food consumption score data}

\item{x}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing identifying information that will be linked to the
food consumption score data}

\item{core.columns}{A vector of variable names to include in resulting
data.frame}
}
\value{
A vector of recoded food consumption score indicators
}
\description{
Function to recode Food Consumption Score for the Myanmar MCCT
Programme Evaluation Survey
}
\examples{
recode_fcs(df = hh)

}
