% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-recode_chealth.R
\name{recode_fever}
\alias{recode_fever}
\title{Function to recode fever prevalence and treatment-seeking behaviour for
the Myanmar MCCT Programme Evaluation Survey}
\usage{
recode_fever(df)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing child health data}
}
\value{
A vector of recoded fever prevalence and treatment-seeking
  behaviour
}
\description{
Function to recode fever prevalence and treatment-seeking behaviour for
the Myanmar MCCT Programme Evaluation Survey
}
\examples{
chealth <- create_chealth(df = childHealth, x = hh, y = hhMembers)
recode_fever(df = chealth)

}
