% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-recode_fplanning.R
\name{recode_fplan}
\alias{recode_fplan}
\title{Function to recode family planning indicators for the Myanmar MCCT
Programme Evaluation Survey}
\usage{
recode_fplan(
  df,
  core.columns = c("KEY", "geo_state", "geo_rural", "geo_villward", "sample_component")
)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing family planning data}

\item{core.columns}{A vector of variable names to include in resulting
data.frame}
}
\value{
A vector of recoded family planning indicators
}
\description{
Function to recode family planning indicators for the Myanmar MCCT
Programme Evaluation Survey
}
\examples{
recode_fplan(df = hh)

}
