% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-recode_wash.R
\name{recode_handwashing}
\alias{recode_handwashing}
\title{Function to recode handwashing indicators for WASH for the Myanmar MCCT
Programme Evaluation Survey}
\usage{
recode_handwashing(df)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing water, sanitation and hygiene data}
}
\value{
A vector of recoded handwashing indicators
}
\description{
Function to recode handwashing indicators for WASH for the Myanmar MCCT
Programme Evaluation Survey
}
\examples{
recode_handwashing(df = hh)

}
