% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-recode_anthro.R
\name{recode_height}
\alias{recode_height}
\title{Function to recode height variable}
\usage{
recode_height(df)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing anthropometric data}
}
\value{
A vector of recoded heights based on any re-testing done
}
\description{
Function to recode height variable
}
\examples{

recode_height(df = childAnthro)

}
