% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-recode_fes.R
\name{recode_hfes}
\alias{recode_hfes}
\title{Function to recode household food expenditure share for the Myanmar MCCT
Programme Evaluation Survey}
\usage{
recode_hfes(
  df,
  core.columns = c("KEY", "geo_state", "geo_rural", "geo_villward", "sample_component")
)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing household food expenditure data}

\item{core.columns}{A vector of variable names to include in resulting
data.frame}
}
\value{
A vector of recoded household food expenditure share indicators
}
\description{
Function to recode household food expenditure share for the Myanmar MCCT
Programme Evaluation Survey
}
\examples{
recode_hfes(df = hh)

}
