% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-recode_iycf.R
\name{recode_iycf}
\alias{recode_iycf}
\title{Function to recode relevant IYCF indicators for the Myanmar MCCT Programme
evaluation study}
\usage{
recode_iycf(
  df,
  core.columns = c("KEY", "geo_state", "geo_rural", "geo_villward", "sample_component",
    "hh_mem_sex")
)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing IYCF/breastfeeding data}

\item{core.columns}{A vector of variable names to include in resulting
data.frame}
}
\value{
A data.frame of recoded IYCF indicators
}
\description{
Function to recode relevant IYCF indicators for the Myanmar MCCT Programme
evaluation study
}
\examples{
iycfDF <- create_iycf(df = iycf, x = hh, y = hhMembers)
recode_iycf(df = iycfDF)

}
