% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-recode_iycf.R
\name{recode_mad}
\alias{recode_mad}
\title{Function to recode minimum acceptable diet indicator for the Myanmar MCCT
Programme Evaluation Survey}
\usage{
recode_mad(df)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing IYCF/breastfeeding data}
}
\value{
A vector of recoded minimum acceptable diet status
}
\description{
Function to recode minimum acceptable diet indicator for the Myanmar MCCT
Programme Evaluation Survey
}
\examples{
iycfDF <- create_iycf(df = iycf, x = hh, y = hhMembers)
recode_mad(df = iycfDF)

}
