% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-recode_anthro.R
\name{recode_maternal_anthro}
\alias{recode_maternal_anthro}
\title{Function to recode maternal MUAC indicators}
\usage{
recode_maternal_anthro(
  df,
  x,
  y,
  core.columns = c("dob", "geo_state", "geo_rural", "geo_villward")
)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing maternal mid-upper arm circumference data}

\item{x}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing identifying information that will be linked to the
mother anthropometry data}

\item{y}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing child anthropometry data to get associated information
that will be linked to the mother anthropometry data}

\item{core.columns}{A vector of variable names to include in resulting}
}
\value{
A data.frame of recoded MUAC measurements
}
\description{
Function to recode maternal MUAC indicators
}
\examples{

recode_maternal_anthro(df = motherAnthro, x = anthroDF, y = childAnthro)

}
