% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19-recode_mcct.R
\name{recode_mcct}
\alias{recode_mcct}
\title{Function to recode MCCT-related indicators for the Myanmar MCCT
Programme Evaluation survey}
\usage{
recode_mcct(
  df,
  core.columns = c("KEY", "geo_rural", "geo_state", "geo_villward", "sample_component")
)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing MCCT-related data}

\item{core.columns}{A vector of variable names to include in resulting
data.frame}
}
\value{
A data.frame of recoded MCCT-related indicators
}
\description{
Function to recode MCCT-related indicators for the Myanmar MCCT
Programme Evaluation survey
}
\examples{
## Recode MCCT-related indicators
mcctDF <- create_mcct(df = hh, x = hhMembers)
recode_mcct(df = mcctDF)

}
