% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-recode_iycf.R
\name{recode_mdd}
\alias{recode_mdd}
\title{Function to recode child food groups for the Myanmar MCCT Programme
Evaluation Survey}
\usage{
recode_mdd(df)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing IYCF/breastfeeding data}
}
\value{
A vector of recoded food groups
}
\description{
Function to recode child food groups for the Myanmar MCCT Programme
Evaluation Survey
}
\examples{
iycfDF <- create_iycf(df = iycf, x = hh, y = hhMembers)
recode_mdd(df = iycfDF)

}
