% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-recode_mddw.R
\name{recode_mddw}
\alias{recode_mddw}
\title{Function to recode minimum dietary diversity for women for the Myanmar MCCT
Programme Evaluation Survey}
\usage{
recode_mddw(
  df,
  core.columns = c("KEY", "geo_state", "geo_rural", "geo_villward", "sample_component")
)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing minimum dietary diversity for women data}

\item{core.columns}{A vector of variable names to include in resulting
data.frame}
}
\value{
A vector of recoded minimum dietary diversity for women indicators
}
\description{
Function to recode minimum dietary diversity for women for the Myanmar MCCT
Programme Evaluation Survey
}
\examples{
recode_mddw(df = hh)

}
