% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-recode_anthro.R
\name{recode_muac}
\alias{recode_muac}
\title{Function to recode MUAC variable}
\usage{
recode_muac(df)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing anthropometric data}
}
\value{
A vector of recoded MUACs based on any re-testing done
}
\description{
Function to recode MUAC variable
}
\examples{

recode_muac(df = childAnthro)

}
