% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18-recode_nbc.R
\name{recode_nbc}
\alias{recode_nbc}
\title{Function to recode newborn care indicators for the Myanmar MCCT
Programme Evaluation survey}
\usage{
recode_nbc(
  df,
  core.columns = c("KEY", "geo_rural", "geo_state", "geo_villward", "sample_component")
)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing newborn care data}

\item{core.columns}{A vector of variable names to include in resulting
data.frame}
}
\value{
A data.frame of recoded newborn care indicators
}
\description{
Function to recode newborn care indicators for the Myanmar MCCT
Programme Evaluation survey
}
\examples{
## Recode newborn care indicators
x <- create_anc(df = anc2, x = hh, y = hhMembers, status = "past")
recode_nbc(df = x)

}
