% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-recode_anthro.R
\name{recode_oedema}
\alias{recode_oedema}
\title{Function to recode oedema variable}
\usage{
recode_oedema(df)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing anthropometric data}
}
\value{
A vector of recoded oedema results based on any re-testing done
}
\description{
Function to recode oedema variable
}
\examples{

recode_oedema(df = childAnthro)

}
