% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17-recode_pnc.R
\name{recode_pnc}
\alias{recode_pnc}
\title{Function to recode postnatal care indicators for the Myanmar MCCT
Programme Evaluation survey}
\usage{
recode_pnc(
  df,
  core.columns = c("KEY", "geo_rural", "geo_state", "geo_villward", "sample_component")
)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing postnatal care data}

\item{core.columns}{A vector of variable names to include in resulting
data.frame}
}
\value{
A data.frame of recoded postnatal care indicators
}
\description{
Function to recode postnatal care indicators for the Myanmar MCCT
Programme Evaluation survey
}
\examples{
## Recode postnatal care indicators
x <- create_anc(df = anc2, x = hh, y = hhMembers, status = "past")
recode_pnc(df = x)

}
