% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-recode_anthro.R
\name{recode_position}
\alias{recode_position}
\title{Function to recode position of height measurements variable}
\usage{
recode_position(df)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing anthropometric data}
}
\value{
A vector of recoded position of height measurements based on any
re-testing done
}
\description{
Function to recode position of height measurements variable
}
\examples{

recode_position(df = childAnthro)

}
