% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-recode_ppi.R
\name{recode_ppi}
\alias{recode_ppi}
\title{Function to recode poverty probability index indicators for the Myanmar MCCT
Programme Evaluation survey}
\usage{
recode_ppi(df, core.columns = c("geo_rural", "geo_state", "sample_component"))
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing poverty-related household data}

\item{core.columns}{A vector of variable names to include in resulting
data.frame}
}
\value{
A data.frame of recoded poverty probability index indicators based on
  the national poverty line
}
\description{
Function to recode poverty probability index indicators for the Myanmar MCCT
Programme Evaluation survey
}
\examples{
## Recode poverty probability index indicators
recode_ppi(df = hh)


}
