% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-recode_anc.R
\name{recode_provider_doctor}
\alias{recode_provider_doctor}
\title{Function to recode the doctor ANC provider data for the Myanmar MCCT
Programme Evaluation survey}
\usage{
recode_provider_doctor(df, status = NULL)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing antenatal care data}

\item{status}{A character value indicating whether to recode data of
currently pregnant women ("current") or previously pregnant women ("past")}
}
\value{
A data.frame of recoded doctor ANC provider information
}
\description{
Function to recode the doctor ANC provider data for the Myanmar MCCT
Programme Evaluation survey
}
\examples{
## Recode provider doctor for currently pregnant women
x <- create_anc(df = anc1, x = hh, y = hhMembers, status = "current")
recode_provider_doctor(df = x, status = "current")

## Recode provider doctor for non-pregnant women
x <- create_anc(df = anc2, x = hh, y = hhMembers, status = "past")
recode_provider_doctor(df = x, status = "past")

}
