% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-recode_anc.R
\name{recode_provider_ehw}
\alias{recode_provider_ehw}
\title{Function to recode the ethnic health worker ANC provider data for the Myanmar
MCCT Programme Evaluation survey}
\usage{
recode_provider_ehw(df, status = NULL)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing antenatal care data for currently pregnant women}

\item{status}{A character value indicating whether to recode data of
currently pregnant women ("current") or previously pregnant women ("past")}
}
\value{
A data.frame of recoded ethnic health worker ANC provider information
}
\description{
Function to recode the ethnic health worker ANC provider data for the Myanmar
MCCT Programme Evaluation survey
}
\examples{
## Recode ehw provider for currently pregnant women
x <- create_anc(df = anc1, x = hh, y = hhMembers, status = "current")
recode_provider_ehw(df = x, status = "current")

## Recode ehw provider for non-pregnant women
x <- create_anc(df = anc2, x = hh, y = hhMembers, status = "past")
recode_provider_ehw(df = x, status = "past")

}
