% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-recode_wash.R
\name{recode_sanitation}
\alias{recode_sanitation}
\title{Function to recode sanitation indicators for WASH for the Myanmar MCCT
Programme Evaluation Survey}
\usage{
recode_sanitation(df)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing water, sanitation and hygiene data}
}
\value{
A vector of recoded sanitation indicators
}
\description{
Function to recode sanitation indicators for WASH for the Myanmar MCCT
Programme Evaluation Survey
}
\examples{
recode_sanitation(df = hh)

}
