% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-recode_anthro.R
\name{recode_sex}
\alias{recode_sex}
\title{Function to recode sex variable}
\usage{
recode_sex(df)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing anthropometric data}
}
\value{
A vector of recoded sex based on any re-testing done
}
\description{
Function to recode sex variable
}
\examples{

recode_sex(df = childAnthro)

}
