% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/26-township.R
\name{recode_township}
\alias{recode_township}
\title{Function to recode township data}
\usage{
recode_township(df)
}
\arguments{
\item{df}{Dataset containing data collected from townships included in
the MCCT baseline study.}
}
\value{
A recoded data.frame with the same number of rows as \code{data}.
}
\description{
Function to recode township data
}
\examples{
recode_township(df = townshipData)

}
