% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/27-village.R
\name{recode_village}
\alias{recode_village}
\title{Function to recode village data}
\usage{
recode_village(df)
}
\arguments{
\item{df}{A data.frame containing village data for the MCCT baseline study}
}
\value{
A data.frame containing recoded variables for village information
}
\description{
Function to recode village data
}
\examples{
recode_village(df = villageData1)

}
