% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-recode_wash.R
\name{recode_wash}
\alias{recode_wash}
\title{Function to recode water, sanitation and hygiene (WASH) indicators for the
Myanmar MCCT Programme Evaluation Survey}
\usage{
recode_wash(
  df,
  core.columns = c("KEY", "geo_state", "geo_rural", "geo_villward", "sample_component")
)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing water, sanitation and hygiene data}

\item{core.columns}{A vector of variable names to include in resulting
data.frame}
}
\value{
A vector of recoded water, sanitation and hygiene indicators
}
\description{
Function to recode water, sanitation and hygiene (WASH) indicators for the
Myanmar MCCT Programme Evaluation Survey
}
\examples{
recode_wash(df = hh)

}
