% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-recode_wash.R
\name{recode_water}
\alias{recode_water}
\title{Function to recode water indicators for WASH for the Myanmar MCCT Programme
Evaluation Survey}
\usage{
recode_water(df)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing water, sanitation and hygiene data}
}
\value{
A vector of recoded water indicators
}
\description{
Function to recode water indicators for WASH for the Myanmar MCCT Programme
Evaluation Survey
}
\examples{
recode_water(df = hh)

}
