% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-recode_anthro.R
\name{recode_weight}
\alias{recode_weight}
\title{Function to recode weight variable}
\usage{
recode_weight(df)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing anthropometric data}
}
\value{
A vector of recoded weights based on any re-testing done
}
\description{
Function to recode weight variable
}
\examples{

recode_weight(df = childAnthro)

}
