% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/25-split_to_quintiles.R
\name{split_to_quintiles}
\alias{split_to_quintiles}
\title{Split observations into quintiles based on poverty probability}
\usage{
split_to_quintiles(df, by = NULL, add = TRUE)
}
\arguments{
\item{df}{A data.frame collected for the Myanmar MCCT Programme Evaluation
Study containing a variable for poverty probability index}

\item{by}{Character value specifying variable in \code{df} to group the
data before splitting into quintiles. Default is NULL and whole data.frame
is split into quintiles. If specified, \code{df} will first be split based
on \code{by} and then each group is split into quintiles}

\item{add}{Logical. Should wealth quintile vector be added to \code{df}? If
TRUE (default), a new variable named \code{wealthQuintile} will be added
to \code{df}. If FALSE, a vector coding of wealth quintiles corresponding
to \code{df}}
}
\value{
If \code{add} is TRUE, returns \code{df} with an added variable
  named \code{wealthQuintile}. If \code{add} is FALSE, returns a vector
  codng of wealth quintiles corresponding to \code{df}
}
\description{
Split observations into quintiles based on poverty probability
}
\examples{
ppiDF <- recode_ppi(df = hh)
split_to_quintiles(df = ppiDF)

}
