% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{townshipData}
\alias{townshipData}
\title{Raw township profile dataset from the Myanmar MCCT Programme Evaluation
Survey}
\format{A data.frame with 70 columns and 7 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
townshipData
}
\description{
Raw township profile dataset from the Myanmar MCCT Programme Evaluation
Survey
}
\keyword{datasets}
