% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{townshipList}
\alias{townshipList}
\title{List of townships included in the study}
\format{A data.frame with 4 columns and 14 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
townshipList
}
\description{
List of townships included in the study
}
\keyword{datasets}
