% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{villageData}
\alias{villageData}
\title{Raw village profile dataset from the Myanmar MCCT Programme Evaluation
Survey}
\format{A data.frame with 3101 columns and 433 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
villageData
}
\description{
Raw village profile dataset from the Myanmar MCCT Programme Evaluation
Survey
}
\keyword{datasets}
