% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{villageData1}
\alias{villageData1}
\title{Raw village profile dataset from the Myanmar MCCT Programme Evaluation
Survey}
\format{A data.frame with 1017 columns and 190 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
villageData1
}
\description{
Raw village profile dataset from the Myanmar MCCT Programme Evaluation
Survey
}
\keyword{datasets}
