% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{villageData2}
\alias{villageData2}
\title{Raw village profile dataset from the Myanmar MCCT Programme Evaluation
Survey}
\format{A data.frame with 1079 columns and 190 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
villageData2
}
\description{
Raw village profile dataset from the Myanmar MCCT Programme Evaluation
Survey
}
\keyword{datasets}
