% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{villageData3}
\alias{villageData3}
\title{Raw village profile dataset from the Myanmar MCCT Programme Evaluation
Survey}
\format{A data.frame with 1016 columns and 170 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
villageData3
}
\description{
Raw village profile dataset from the Myanmar MCCT Programme Evaluation
Survey
}
\keyword{datasets}
