% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{washResults}
\alias{washResults}
\title{WASH indicators results table}
\format{A data.frame with 9 columns and 180 rows.}
\source{
Myanmar MCCT Programme Evaluation Survey
}
\usage{
washResults
}
\description{
WASH indicators results table
}
\keyword{datasets}
